third_party/blink/web_tests/external/wpt/svg/styling/attr-style-type-expected.txt

This is a testharness.js-based test.
[FAIL] type attribute on SVG <style> elements should reflect correctly
  assert_equals: missing type reflects as empty string IDL attribute on the style element expected "" but got "text/css"
Harness: the test ran to completion.
